remove FIFOSet and update ut
