+++ src/neo/Helper.cs
@@ -83,19 +83,6 @@ namespace Neo
             }
         }
 
-        internal static void Remove<T>(this HashSet<T> set, FIFOSet<T> other)
-            where T : IEquatable<T>
-        {
-            if (set.Count > other.Count)
-            {
-                set.ExceptWith(other);
-            }
-            else
-            {
-                set.RemoveWhere(u => other.Contains(u));
-            }
-        }
-
         internal static void Remove<T>(this HashSet<T> set, HashSetCache<T> other)
             where T : IEquatable<T>
         {
