+++ tests/neo.UnitTests/IO/Caching/UT_HashSetCache.cs
@@ -1,6 +1,9 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.IO.Caching;
+using System;
+using System.Collections;
+using System.Linq;
 
 namespace Neo.UnitTests.IO.Caching
 {
@@ -39,6 +42,116 @@ namespace Neo.UnitTests.IO.Caching
 
             bucket.Contains(13).Should().BeFalse();
             bucket.Contains(50).Should().BeTrue();
+        }
+
+        [TestMethod]
+        public void TestConstructor()
+        {
+            Action action1 = () => new HashSetCache<UInt256>(-1);
+            action1.Should().Throw<ArgumentOutOfRangeException>();
+
+            Action action2 = () => new HashSetCache<UInt256>(1, -1);
+            action2.Should().Throw<ArgumentOutOfRangeException>();
+
+            Action action3 = () => new HashSetCache<UInt256>(1, 21);
+            action3.Should().Throw<ArgumentOutOfRangeException>();
+        }
+
+        [TestMethod]
+        public void TestAdd()
+        {
+            var a = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01
+            });
+            var b = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x02
+            });
+            var set = new HashSetCache<UInt256>(1, 1)
+            {
+                a,
+                b
+            };
+            CollectionAssert.AreEqual(set.ToArray(), new UInt256[] { b });
+        }
+
+        [TestMethod]
+        public void TestGetEnumerator()
+        {
+            var a = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01
+            });
+            var b = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x02
+            });
+            var set = new HashSetCache<UInt256>(1, 1)
+            {
+                a,
+                b
+            };
+            IEnumerable ie = set;
+            ie.GetEnumerator().Should().NotBeNull();
+        }
+
+        [TestMethod]
+        public void TestExceptWith()
+        {
+            var a = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01
+            });
+            var b = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x02
+            });
+            var c = new UInt256(new byte[32] {
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
+                0x01, 0x03
+            });
+
+            var set = new HashSetCache<UInt256>(10)
+            {
+                a,
+                b,
+                c
+            };
+            set.ExceptWith(new UInt256[] { b, c });
+            CollectionAssert.AreEqual(set.ToArray(), new UInt256[] { a });
+
+            set = new HashSetCache<UInt256>(10)
+            {
+                a,
+                b,
+                c
+            };
+            set.ExceptWith(new UInt256[] { a });
+            CollectionAssert.AreEqual(set.ToArray(), new UInt256[] { b, c });
+
+            set = new HashSetCache<UInt256>(10)
+            {
+                a,
+                b,
+                c
+            };
+            set.ExceptWith(new UInt256[] { c });
+            CollectionAssert.AreEqual(set.ToArray(), new UInt256[] { a, b });
         }
     }
 }
